Implementirane metode za dobavljanje izvjestaja o poslovanju klinike.
	Implementirana metoda getRevenue za vracanje prihoda u odredjenom
vremenskom periodu koji administrator unosi kao dva datuma.
	Implementirana metoda getReportForWeek za prikaz broja operacija i
pregleda u toku sedam prethodnih dana na grafiku. Metoda vraca broj
operacija i pregleda za sedam prethodnih dana.
	Implementirana metoda getReportForMonth za prikaz broja operacija i
pregleda po mjesecima u godini. Metoda vraca broj pregleda i operacija
za svaki mjesec u toku godine.

Implementirane metode za odbijanje i prihvatanje zahtijeva za odsustvo
ili godisnji odmor.
Izvrsene modifikacije u modelu (klasa Absence), dodata dodatna polja.
Implementirana metoda za slanje email-a medicinskom radniku u kome se
navodi razlog odbijanja, u koliko je zahtjev odbijen, ili samo
obavijestenje da je zahtjev prihvacen.

Izvrsene odgovarajuce izmjene za potrebe ovih funkcionalnosti.

+++ psw-isa/src/main/java/tim31/pswisa/dto/CheckUpTypeDTO.java
@@ -6,13 +6,13 @@ public class CheckUpTypeDTO {
 
 	private Long id;
 	private String name;
-	private int typePrice;
+	private double typePrice;
 
 	public CheckUpTypeDTO(CheckUpType c) {
 		this(c.getId(), c.getName(), c.getTypePrice());
 	}
 
-	public CheckUpTypeDTO(Long id, String name, int price) {
+	public CheckUpTypeDTO(Long id, String name, double price) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -32,11 +32,11 @@ public class CheckUpTypeDTO {
 		this.id = id;
 	}
 
-	public int getTypePrice() {
+	public double getTypePrice() {
 		return typePrice;
 	}
 
-	public void setTypePrice(int typePrice) {
+	public void setTypePrice(double typePrice) {
 		this.typePrice = typePrice;
 	}
